llama workload multi server fixv0

+++ ais-bench_workload/src/train/huawei/train_mindspore_llama/scripts/cluster_offline_run.sh
@@ -50,15 +50,30 @@ init()
 run_train()
 {
     logger_Info "-------------------------------- train start --------------------------------"
-    cmd="export WORK_PATH=$WORK_PATH;
+    env_cmd=cmd="export WORK_PATH=$WORK_PATH;
         export RESULT_PATH=$RESULT_PATH;
         export PYTHONPATH=$WORK_PATH:$PYTHONPATH;
         export PYTHONPATH=$WORK_PATH/logging:$PYTHONPATH;
-        source $WORK_PATH/config/$CONFIG_FILE;
-        rm -rf $RESULT_PATH/*.json;
-        bash $WORK_PATH/run_node.sh train"
+        source $WORK_PATH/config/$CONFIG_FILE"
 
-    cluster_run_cmd_parallel "${NODEINFO_FILE}" ${cmd} || { logger_Warn "run train failed"; return 1; }
+    if [ "$LLAMA_RUN_MODE" == "full" ] && [ "$LLAMA_RUN_MODE" == "only_pretrain" ];then
+        cmd="$env_cmd;
+            rm -rf $RESULT_PATH/*.json;
+            bash $WORK_PATH/run_node.sh train train"
+        cluster_run_cmd_parallel "${NODEINFO_FILE}" ${cmd} || { logger_Warn "run train(pretrain) failed"; return 1; }
+        cluster_rscp "${NODEINFO_FILE}" ${RESULT_PATH} ${RESULT_PATH}
+        $env_cmd
+        bash $WORK_PATH/run_node.sh merge || { logger_Warn "ckpt merge failed"; return 1; }
+    fi
+    if [ "$LLAMA_RUN_MODE" == "full" ] && [ "$LLAMA_RUN_MODE" == "only_finetune" ];then
+        cmd="$env_cmd;
+            rm -rf $RESULT_PATH/*.json;
+            bash $WORK_PATH/run_node.sh train finetune"
+        cluster_run_cmd_parallel "${NODEINFO_FILE}" ${cmd} || { logger_Warn "run train(finetune) failed"; return 1; }
+        cluster_rscp "${NODEINFO_FILE}" ${RESULT_PATH} ${RESULT_PATH}
+        $env_cmd
+        bash $WORK_PATH/run_node.sh merge || { logger_Warn "ckpt merge failed"; return 1; }
+    fi
     logger_Info "-------------------------------- train end --------------------------------"
 }
